backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sentence-transformers
faiss-cpu
numpy
PyPDF2
openai
python-dotenv